fix: prevent chrome's click-and-drag behavior from closing modals
The web-standard dictates that, when the mouse button is held down and
the mouse is dragged over multiple elements, the click event is fired on
the closest common ancestor between start and end elements when the
mouse button is lifted.

Firefox doesn't adhere to this standard for form fields and instead does
not fire a click event, but other browsers do that, thus closing the
form when a user drags the mouse over a form field (for example when
selecting text) and out of the modal dialog into the splash.